feat: header component 분리 및 페이지 적용
